Fixes LIBTD-110: Hide password entry from displaying on the screen when prompt for an admin user
Resource type for collection will be Dataset only and for generic file will be one of the seven options

Revert the changes in rake tasks as it is unrelated to other changes

When creating a new admin user with rake task, should not echo the password on the screen. Fix LIBTD-110.

Set up a default admin user's email and password instead of prompting for a random one

Use find_or_create_by to concisely create the nonexisting default roles