Configure e-mail settings in Data Repository application.
Configure Action Mailer to send e-mail in production but not in
development.  Action Mailer is set to use the sendmail transport.

We also set the address to send mail to in the Contact form and the
e-mail address the contact form appears to come from, too.

For now, root@localhost is being used as a placeholder address in
production.rb, until a permanent delivery address is obtained.

Note: the ContactForm configuration options were rather obscure to
discern.