Update to Rails 4.2.5.1
Update Rails to 4.2.5.1 in light of recent security advisories.  Note,
we keep the sprockets and sprockets-rails gems at their former values
for now because the latest versions appear not to work with our version
of Sufia.

Because of the update, the step in the Travis build that creates the
config/secrets.yml file is complaining about rake:

rake aborted!
Gem::LoadError: You have already activated rake 10.4.2, but your
Gemfile requires rake 10.5.0. Prepending `bundle exec` to your command
may solve this.

We fix the rake conflict by following the above advice. :-)

Also updated the version of Rails used in the testing matrix to 4.2.5.1

Added a note in the Gemfile explaining why we are specifying the
older versions of the sprockets gems.