Backport improvements from ETDplus development
These changes are in tandem with the changes to the InstallScripts
repository that incorporate improvements from development on the
ETDplus project.

Because we are no longer doing a "rails new" during install, we must
ensure we have certain necessary directories.  The existing .gitignore
means we do not have a log directory when cloning the application code.
 This causes rake tasks to fail because they can't log within their
RAILS_ENV environment.