Comment sample secrets.yml file
(As from the install scripts:)
Add comments to the secrets.yml.sample file that is intended to be
edited and dropped into files/ during bootstrap. The comments give
some guidance as to the meanings of the various YAML keys. Keys that
are not normally present are included but commented out.

Also add some extra anchors to the default section so that
individual items may be overridden, like that done for fedora:. These
anchors are also included in the vanilla secrets.yml file generated if
no files/secrets.yml file is present when running bootstrap.sh.