Update homepage carousel.
Add new marketing text to the carousel, and load a random image from
/app/assets/images/carousel for the background image. Added a rake task
to add photos from a zip to the folder. It takes a parameter that is the
path to the zip file. E.g:

bundle exec rake datarepo:update_images['/path/to/images.zip']

The task will delete the images currently in the folder.