Add additional settings to define which ORCiD service is used
The ORCiD service has available a sandbox and production service.  The
one used is determined by various URLs assigned to ORCID_*_URL
environment variables, which define API endpoints.  If these
environment variables are unset then they default to the ones used by
the sandbox service.

This change extends the settings that may be assigned in
config/application.yml.  The "orcid" gem uses this file to set
ORCiD-related environment variables.  We, in turn, derive the actual
value set from the "orcid" key in config/secrets.yml.

This will allow the application to use the sandbox or production ORCiD
service.