Use forked version of resque gem
The "resque-pool" service currently fails to start up except in the Rails "production" environment.  This is due to a failure of the "resque" gem to initialise due to a "Cannot define multiple 'included' blocks for a Concern" error.  This problem was fixed in the "resque" gem in pull request 1597 (https://github.com/resque/resque/pull/1597), but has not been incorporated into an official gem release yet.

This change alters "Gemfile" to depend upon a forked version of the last "1.x" release of the "resque" gem that incorporates the cherry-picked fix for the aforementioned problem.

Also included is the "Gemfile.lock" resulting from a "bundle install" resulting from the new "Gemfile" dependencies.